New map organisation
The elevators are now working as portals and send the player to the new position.
Because of this, the 2. floor is now added to the main tilemap.
It's just placed far away, so that it's not possible to see it at the beginning.
works for just this little level.

Old scrips and second TileMapLayer should be all deleted now. we don't need them anymore.